requirements.txt
requests
pandas
python-dateutil
matplotlib
seaborn
plotly
pytest